+++ pyproject.toml
@@ -72,7 +72,7 @@ dev = [
     "necessary",
     "peft",
     "datasets",
-    "omegaconf",
+    "omegaconf"
 ]
